requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-dotenv==0.10.1
six==1.12.0
# Editable Git install with no remote (website==0.1.0)
-e /home/hasone/code/nas_website
Werkzeug==0.15.1
